[readability] move responser formatting logic to separate module
